backend/pyproject.toml
[tool.poetry]
name = "toeic-exercise-backend"
version = "0.1.0"
description = "TOEIC Grammar Practice - FastAPI Backend"
authors = ["Developer"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {version = "^0.115.0", extras = ["standard"]}
uvicorn = {version = "^0.32.0", extras = ["standard"]}
sqlalchemy = {version = "^2.0.0", extras = ["asyncio"]}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic-settings = "^2.6.0"
google-generativeai = "^0.8.0"
python-dotenv = "^1.0.0"
greenlet = "^3.3.1"
groq = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
